+++ lib/html.dart
@@ -7,6 +7,7 @@ library rikulo_html;
 import "dart:html";
 import "dart:math";
 
+import "package:meta/meta.dart";
 import 'package:rikulo_commons/util.dart';
 
 part "src/html/Browser.dart";